feat(bot): install redis & upgrade FSM as storage
