+++ backend/conf/settings.py
@@ -112,7 +112,7 @@ CHANNEL_LAYERS = {
         "CONFIG": {
             "hosts": [
                 (os.getenv('REDIS_HOST', default='localhost'),
-                 os.getenv('REDIS_PORT', default='6379'))
+                 os.getenv('REDIS_PORT', default='6379') + '/0')
             ],
         },
     }
@@ -297,8 +297,8 @@ if DEV_SERVICES:
 CELERY_TIMEZONE = TIME_ZONE
 CELERY_TASK_SERIALIZER = 'pickle'
 CELERY_ACCEPT_CONTENT = ('pickle',)
-CELERY_BROKER_URL = os.getenv('CELERY_BROKER', 'redis://redis:6379')
-CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT', 'redis://redis:6379')
+CELERY_BROKER_URL = os.getenv('CELERY_BROKER', 'redis://redis:6379/1')
+CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT', 'redis://redis:6379/2')
 CELERY_BEAT_MAX_LOOP_INTERVAL = 20
 CELERY_BEAT_SCHEDULER = 'django_celery_beat.schedulers:DatabaseScheduler'
 CELERY_BEAT_SCHEDULE = {
