+++ bot/config_reader.py
@@ -16,10 +16,12 @@ class Settings(BaseSettings):
     WEB_HOOK_MODE: bool
     WEB_HOOK_HOST: HttpUrl
     WEB_APP_PORT: int
+    REDIS_HOST: str
+    REDIS_PORT: int
 
     @property
     def get_web_hook_url(self) -> HttpUrl:
-        return urljoin(self.WEB_HOOK_HOST + 'telegram_webhook/')
+        return urljoin(self.WEB_HOOK_HOST, 'telegram_webhook/')
 
     @property
     def get_timezone(self) -> tzinfo:
