+++ bot/pyproject.toml
@@ -6,7 +6,6 @@ authors = ["Moodbeat company"]
 
 [tool.poetry.dependencies]
 python = "^3.9"
-pre-commit = "3.3.2"
 aiogram = "3.0.0b7"
 SQLAlchemy = "2.0.15"
 aiosqlite = "0.19.0"
@@ -14,13 +13,14 @@ python-dotenv = "1.0.0"
 email-validator = "2.0.0.post2"
 requests = "2.31.0"
 pytz = "^2023.3"
+redis = "^4.6.0"
 
 
 [tool.poetry.group.dev.dependencies]
-django-debug-toolbar = "4.0.0"
 autopep8 = "2.0.2"
 flake8 = "6.0.0"
 isort = "^5.12.0"
+pre-commit = "^3.3.3"
 
 [build-system]
 requires = ["poetry-core"]
