+++ bot/services/condition_service.py
@@ -1,3 +1,5 @@
+from urllib.parse import urljoin
+
 from aiogram.fsm.context import FSMContext
 
 from config_reader import config
@@ -12,7 +14,7 @@ async def get_current_user_with_condition(
 ) -> UserConditionGetResponse:
     headers = await get_headers(state)
     data = await make_get_request(
-        config.BASE_ENDPOINT + 'users/current_user/',
+        urljoin(config.BASE_ENDPOINT, 'users/current_user/'),
         headers=headers
     )
     return UserConditionGetResponse(**data)
@@ -25,7 +27,7 @@ async def post_condition_data(
     data = ConditionPostRequest(mood=mood)
     headers = await get_headers(state)
     await make_post_request(
-        config.BASE_ENDPOINT + 'metrics/conditions/',
+        urljoin(config.BASE_ENDPOINT, 'metrics/conditions/'),
         data=data.dict(),
         headers=headers
     )
