+++ bot/services/user_service.py
@@ -1,3 +1,5 @@
+from urllib.parse import urljoin
+
 from aiogram.fsm.context import FSMContext
 
 from config_reader import config
@@ -9,7 +11,7 @@ from .api.response_models import CurrentUserGetResponse
 async def get_current_user(state: FSMContext) -> CurrentUserGetResponse:
     headers = await get_headers(state)
     data = await make_get_request(
-        config.BASE_ENDPOINT + 'users/current_user/',
+        urljoin(config.BASE_ENDPOINT, 'users/current_user/'),
         headers=headers
     )
     return CurrentUserGetResponse(**data)
